feat: Add loading screen
On chrome there is the issues when no user interaction happened
yet that it will not allow the AudioContext to start which results
in queueing sounds until it is started. Loading the page and
spamming ESC will queue many sounds and then play a very loud sound
if the user interacts with the website. Adding this screen forces
the user to interact with the website and resolves this issue